+++ doc/release_notes.md
@@ -8,6 +8,10 @@
 
 Please list contributions, add reference to PRs if present.
 
+* Integrated **demand redistribution** based on utility and balancing authority level demand [PR #34](https://github.com/open-energy-transition/efuels-supply-potentials/pull/34)
+
+* Updated **custom_powerplants.csv** data with new entries for coal and ror from EIA [PR #29](https://github.com/open-energy-transition/efuels-supply-potentials/pull/29)
+
 * Enabled **custom data moving** within workflow (`custom_powerplants.csv` and `NorthAmerica.csv`): [PR #21](https://github.com/open-energy-transition/efuels-supply-potentials/pull/21)
 
 * Enabled **retrieving precomputed output** of power model's rules (e.g. `download_osm_data`, `clean_osm_data`, `build_shapes`, `build_osm_network`, `base_network`, and `build_renewable_profiles`) within the workflow: [PR #20](https://github.com/open-energy-transition/efuels-supply-potentials/pull/20)